GH-1400 Move metrics publisher auto-configuration to core.metrics
